Refactor todo class
I'm anticipating that we're going to need to start analyzing these lines
more, so I moved the character number analysis to an instance method,
because if we're going to any analysis of the line in this class, we
should do all of the analysis in the class.

Also started to implement a RequiredArg class, but decided to extract
that into its own gem: https://github.com/maxjacobson/required_arg

I'm planning to run todo lint on Ruby 2.0.0, and I want to have required
keyword arguments.